requirements.txt
# Core ML/DL Dependencies
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Handling
pandas>=1.3.0
networkx>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration and Logging
pyyaml>=6.0.0
tqdm>=4.64.0

# Development
pytest>=7.0.0
jupyter>=1.0.0
ipykernel
